Remove `which()`
This parsing library aims to make it simple to write abstract parsers. It seeks to eliminate the need for the user to write custom parsers, and instead provide the building blocks to assemble any parser they desire.
Thus, don't see any use for `which()`, since it's only really useful for manual parsers and would disencourage the assembly of parsers. Especially with the addition of `convert()` I don't really see it as having any use at this point.
Also, the removal simplifies the library, which is great.